LearnApp-API/requirements.txt
fastapi==0.110.0
pydantic==2.11.7
uvicorn==0.29.0
google-genai
websocket-client==1.8.0
